Split up threshold and parser
