+++ src/query/api/v1/handler/prometheus/native/parse_query.go
@@ -26,7 +26,6 @@ import (
 	"strings"
 
 	"github.com/m3db/m3/src/query/api/v1/handler"
-	"github.com/m3db/m3/src/query/functions/scalar"
 	"github.com/m3db/m3/src/query/models"
 	"github.com/m3db/m3/src/query/parser"
 	"github.com/m3db/m3/src/query/parser/promql"
@@ -59,29 +58,6 @@ func NewPromParseHandler(
 	}
 }
 
-// QueryRepresentation is a JSON representation of a query after attempting to
-// extract any threshold-specific parameters.
-//
-// NB: this is always presented with the threshold value (if present) as
-// appearing on the right of the query.
-// e.g. `1 > up` will instead be inverted to `up < 1`
-type QueryRepresentation struct {
-	// Query is the non-threshold part of a query.
-	Query FunctionNode `json:"query,omitempty"`
-	// Threshold is any detected threshold (top level comparison functions).
-	// NB: this is a pointer so it does not have to be present in output if unset.
-	Thresold *Threshold `json:"threshold,omitempty"`
-}
-
-// Threshold is a JSON representation of a threshold, represented by a top level
-// comparison function.
-type Threshold struct {
-	// Comparator is the threshold comparator.
-	Comparator string `json:"comparator"`
-	// Value is the threshold value.
-	Value float64 `json:"value"`
-}
-
 // FunctionNode is a JSON representation of a function.
 type FunctionNode struct {
 	// Name is the name of the function node.
@@ -105,89 +81,6 @@ func (n FunctionNode) String() string {
 	}
 
 	return sb.String()
-}
-
-var thresholdNames = []string{">", ">=", "<", "<="}
-
-type thresholdInfo struct {
-	thresholdIdx int
-	queryIdx     int
-	comparator   string
-}
-
-// isThreshold determines if this function node is a candidate for being a
-// threshold function, and returns the index of the child representing the
-// threshold, and the index of the child for the rest of the query.
-func (n FunctionNode) isThreshold() (bool, thresholdInfo) {
-	info := thresholdInfo{-1, -1, ""}
-	// NB: only root nodes with two children may be thresholds.
-	if len(n.parents) != 0 || len(n.Children) != 2 {
-		return false, info
-	}
-
-	for i, name := range thresholdNames {
-		if n.Name != name {
-			continue
-		}
-
-		scalarLeft := n.Children[0].node.Op.OpType() == scalar.ScalarType
-		scalarRight := n.Children[1].node.Op.OpType() == scalar.ScalarType
-
-		// NB: if both sides are scalars, it's a calculator, not a query.
-		if scalarLeft && scalarRight {
-			return false, info
-		}
-
-		if scalarLeft {
-			// NB: Thresholds are standardized to have the threshold value on the RHS;
-			// if the scalar is on the left, the operation must be inverted.
-			inverted := thresholdNames[(i+2)%len(thresholdNames)]
-			return true, thresholdInfo{
-				thresholdIdx: 0,
-				queryIdx:     1,
-				comparator:   inverted,
-			}
-		}
-
-		if scalarRight {
-			return true, thresholdInfo{
-				thresholdIdx: 1,
-				queryIdx:     0,
-				comparator:   name,
-			}
-		}
-
-		break
-	}
-
-	return false, info
-}
-
-// QueryRepresentation gives the query representation of the function node.
-func (n FunctionNode) QueryRepresentation() (QueryRepresentation, error) {
-	isThreshold, thresholdInfo := n.isThreshold()
-	if !isThreshold {
-		return QueryRepresentation{
-			Query: n,
-		}, nil
-	}
-
-	queryPart := n.Children[thresholdInfo.queryIdx]
-	thresholdPart := n.Children[thresholdInfo.thresholdIdx]
-	thresholdOp := thresholdPart.node.Op
-	scalarOp, ok := thresholdOp.(*scalar.ScalarOp)
-	if !ok {
-		return QueryRepresentation{}, instrument.InvariantErrorf(
-			"operation %s must be a scalar", thresholdOp.String())
-	}
-
-	return QueryRepresentation{
-		Query: queryPart,
-		Thresold: &Threshold{
-			Comparator: thresholdInfo.comparator,
-			Value:      scalarOp.Value(),
-		},
-	}, nil
 }
 
 type nodeMap map[string]FunctionNode
@@ -263,55 +156,54 @@ func parsingThreshold(r *http.Request) bool {
 	return true
 }
 
-func (h *promParseHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	logger := h.instrumentOpts.Logger()
+func parseRootNode(
+	w http.ResponseWriter,
+	r *http.Request,
+	logger *zap.Logger,
+) (FunctionNode, bool) {
 	query, err := parseQuery(r)
 	if err != nil {
 		xhttp.Error(w, err, http.StatusBadRequest)
 		logger.Error("cannot parse query string", zap.Error(err))
-		return
+		return FunctionNode{}, false
 	}
 
 	parser, err := promql.Parse(query, models.NewTagOptions())
 	if err != nil {
 		xhttp.Error(w, err, http.StatusBadRequest)
 		logger.Error("cannot parse query PromQL", zap.Error(err))
-		return
+		return FunctionNode{}, false
 	}
 
 	nodes, edges, err := parser.DAG()
 	if err != nil {
 		xhttp.Error(w, err, http.StatusBadRequest)
 		logger.Error("cannot extract query DAG", zap.Error(err))
-		return
+		return FunctionNode{}, false
 	}
 
 	funcMap, err := constructNodeMap(nodes, edges)
 	if err != nil {
 		xhttp.Error(w, err, http.StatusBadRequest)
 		logger.Error("cannot construct node map", zap.Error(err))
-		return
+		return FunctionNode{}, false
 	}
 
 	root, err := funcMap.rootNode()
 	if err != nil {
 		xhttp.Error(w, err, http.StatusBadRequest)
 		logger.Error("cannot fetch root node", zap.Error(err))
-		return
+		return FunctionNode{}, false
 	}
 
-	// NB: If not parsing threshold, return the raw AST.
-	if !parsingThreshold(r) {
+	return root, true
+}
+
+func (h *promParseHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	logger := h.instrumentOpts.Logger()
+	root, success := parseRootNode(w, r, logger)
+	// NB: if successful, need to write teh response.
+	if success {
 		xhttp.WriteJSONResponse(w, root, logger)
-		return
 	}
-
-	queryRepresentation, err := root.QueryRepresentation()
-	if err != nil {
-		xhttp.Error(w, err, http.StatusBadRequest)
-		logger.Error("cannot convert to query representation", zap.Error(err))
-		return
-	}
-
-	xhttp.WriteJSONResponse(w, queryRepresentation, logger)
 }
